Use Rect2f for RectPrimitive area.
Allegro uses float arguments in al_draw_rect calls, so might as well
store the rect primitive data as floats.

+++ src/cid/graphics/render.d
@@ -221,7 +221,7 @@ struct TextBatch {
 }
 
 struct RectPrimitive {
-  Rect2i   rect;
+  Rect2f   rect;
   bool     centered;
   bool     filled;
   float    thickness = 1f;
